Fix up some remaining windows things and add gitignore for windows/vs Update font detection to read files themselve for info rather than relying on naming

+++ gwen/include/Gwen/TextObject.h
@@ -89,7 +89,7 @@ namespace Gwen
 				return m_Unicode;
 			}
 
-			int length() const { return m_Unicode.length(); }
+			int length() const { return (int)m_Unicode.length(); }
 
 			Gwen::UnicodeString		m_Unicode;
 			Gwen::String			m_String;
